+++ engine/thorium/message.h
@@ -176,8 +176,8 @@ void thorium_message_print_tracepoints(struct thorium_message *self);
 
 void thorium_message_set_identifier(struct thorium_message *self, uint64_t identifier);
 void thorium_message_set_parent_identifier(struct thorium_message *self, uint64_t identifier);
-int thorium_message_get_identifier(struct thorium_message *self);
-int thorium_message_get_parent_identifier(struct thorium_message *self);
+uint64_t thorium_message_get_identifier(struct thorium_message *self);
+uint64_t thorium_message_get_parent_identifier(struct thorium_message *self);
 
 /*
  * Metadata functions.
